Encapsulating the logic in a new DataCleanerHomeUpgrader class.
