+++ desktop/ui/src/main/java/org/datacleaner/user/DataCleanerHome.java
@@ -26,14 +26,8 @@ import java.io.OutputStream;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 
-import org.apache.commons.vfs2.AllFileSelector;
-import org.apache.commons.vfs2.FileDepthSelector;
 import org.apache.commons.vfs2.FileObject;
 import org.apache.commons.vfs2.FileSystemException;
 import org.apache.commons.vfs2.FileSystemManager;
@@ -163,20 +157,10 @@ public final class DataCleanerHome {
         }
 
         if (!isUsable(candidate)) {
-            // First try to find previous versions of DC Home folder and upgrade
-            // instead of starting from scratch
-            FileObject upgradeCandidate = findUpgradeCandidate(candidate);
+            DataCleanerHomeUpgrader upgrader = new DataCleanerHomeUpgrader();
+            FileObject upgradeCandidate = upgrader.upgrade(candidate);
 
-            if (upgradeCandidate != null) {
-                logger.debug("Upgrading DATACLEANER_HOME from : {}", upgradeCandidate);
-                candidate.copyFrom(upgradeCandidate, new AllFileSelector());
-
-                // Overwrite example jobs
-                final List<String> allFilePaths = DemoConfiguration.getAllFilePaths();
-                for (String filePath : allFilePaths) {
-                    copyFile(candidate, manager, filePath, true);
-                }
-            } else {
+            if (upgradeCandidate == null) {
                 logger.debug("Copying default configuration and examples to DATACLEANER_HOME directory: {}", candidate);
                 copyFile(candidate, manager, "conf.xml", false);
 
@@ -187,106 +171,6 @@ public final class DataCleanerHome {
             }
         }
         return candidate;
-    }
-
-    private static FileObject findUpgradeCandidate(FileObject candidate) throws FileSystemException {
-        List<String> candidateBlacklist = Arrays.asList("log", Version.UNKNOWN_VERSION);
-
-        FileObject parent = candidate.getParent();
-        List<FileObject> versionFolders = new ArrayList<>();
-        FileObject[] allFoldersInParent = parent.findFiles(new FileDepthSelector(1, 1));
-        for (FileObject folderInParent : allFoldersInParent) {
-            final String folderInParentName = folderInParent.getName().getBaseName();
-            if (folderInParent.getType().equals(FileType.FOLDER) && (!folderInParent.equals(candidate))
-                    && (!candidateBlacklist.contains(folderInParentName))) {
-                versionFolders.add(folderInParent);
-            }
-        }
-
-        List<FileObject> validateVersionFolders = validateVersionFolders(versionFolders);
-
-        if (!validateVersionFolders.isEmpty()) {
-
-            FileObject latestVersion = Collections.max(validateVersionFolders, new Comparator<FileObject>() {
-
-                @Override
-                public int compare(FileObject o1, FileObject o2) {
-                    String o1BaseName = o1.getName().getBaseName();
-                    String o2BaseName = o2.getName().getBaseName();
-
-                    String[] o1Split = o1BaseName.split("\\.");
-                    String[] o2Split = o2BaseName.split("\\.");
-
-                    for (int i = 0; i < Math.min(o1Split.length, o2Split.length); i++) {
-                        Integer o1Part;
-                        if (o1Split[i].endsWith("-SNAPSHOT")) {
-                            o1Part = Integer.parseInt(o1Split[i].substring(0, o1Split[i].lastIndexOf("-SNAPSHOT")));
-                        } else {
-                            o1Part = Integer.parseInt(o1Split[i]);
-                        }
-                        Integer o2Part;
-                        if (o2Split[i].endsWith("-SNAPSHOT")) {
-                            o2Part = Integer.parseInt(o2Split[i].substring(0, o2Split[i].lastIndexOf("-SNAPSHOT")));
-                        } else {
-                            o2Part = Integer.parseInt(o2Split[i]);
-                        }
-
-                        int compareTo = o1Part.compareTo(o2Part);
-                        if (compareTo == 0) {
-                            // check another part
-                            continue;
-                        } else {
-                            return compareTo;
-                        }
-                    }
-
-                    Integer o1SplitLength = (Integer) o1Split.length;
-                    Integer o2SplitLength = (Integer) o2Split.length;
-                    return o1SplitLength.compareTo(o2SplitLength);
-                }
-            });
-            return latestVersion;
-        } else {
-            return null;
-        }
-    }
-
-    private static List<FileObject> validateVersionFolders(List<FileObject> versionFolders) {
-        List<FileObject> validatedVersionFolders = new ArrayList<>();
-        Integer currentMajorVersion = Version.getMajorVersion();
-        if (currentMajorVersion == null) {
-            return validatedVersionFolders;
-        }
-
-        for (FileObject versionFolder : versionFolders) {
-            String baseName = versionFolder.getName().getBaseName();
-
-            String[] versionParts = baseName.split("\\.");
-
-            try {
-                int majorVersion = Integer.parseInt(versionParts[0]);
-                if (majorVersion != currentMajorVersion) {
-                    continue;
-                }
-            } catch (NumberFormatException e) {
-                continue;
-            }
-
-            for (String versionPart : versionParts) {
-                if (versionPart.endsWith("-SNAPSHOT")) {
-                    versionPart = versionPart.substring(0, versionPart.lastIndexOf("-SNAPSHOT"));
-                }
-                try {
-                    Integer.parseInt(versionPart);
-                    validatedVersionFolders.add(versionFolder);
-                } catch (NumberFormatException e) {
-                    logger.warn(
-                            "Found a version folder in home directory ({}) with a part that could not be parsed to an integer: {} Removing this folder from potential upgrade candidates.",
-                            baseName, versionPart);
-                }
-            }
-        }
-        return validatedVersionFolders;
     }
 
     private static FileObject initializeDataCleanerHomeFallback() throws FileSystemException {
